@@ -1,15 +1,15 @@
 name: flutter_html
 description: A Flutter widget for rendering static html tags as Flutter widgets. (Will render over 70 different html tags!)
-version: 0.11.1
+version: 0.12.0
 author: Matthew Whitaker <sub6resources@gmail.com>
 homepage: https://github.com/Sub6Resources/flutter_html
 
 environment:
-  sdk: '>=1.19.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
   flutter: '>=0.5.0'
 
 dependencies:
-  html: ^0.14.0
+  html: ^0.15.4
   flutter:
     sdk: flutter
 
