@@ -19,9 +19,9 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, this.title}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   _MyHomePageState createState() => new _MyHomePageState();
@@ -32,7 +32,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return new Scaffold(
       appBar: new AppBar(
-        title: new Text(widget.title),
+        title: new Text(widget.title!),
       ),
       body: new Center(
         child: SingleChildScrollView(
@@ -151,14 +151,14 @@ class _MyHomePageState extends State<MyHomePage> {
               }
               return null;
             },
-            customTextStyle: (dom.Node node, TextStyle baseStyle) {
+            customTextStyle: (dom.Node node, TextStyle? baseStyle) {
               if (node is dom.Element) {
                 switch (node.localName) {
                   case "p":
-                    return baseStyle.merge(TextStyle(height: 2, fontSize: 20));
+                    return baseStyle!.merge(TextStyle(height: 2, fontSize: 20));
                 }
               }
-              return baseStyle;
+              return baseStyle ?? TextStyle();
             },
           ),
         ),
