@@ -8,8 +8,8 @@ import 'image_properties.dart';
 
 class Html extends StatelessWidget {
   Html({
-    Key key,
-    @required this.data,
+    Key? key,
+    required this.data,
     this.padding,
     this.backgroundColor,
     this.defaultTextStyle,
@@ -33,32 +33,32 @@ class Html extends StatelessWidget {
   }) : super(key: key);
 
   final String data;
-  final EdgeInsetsGeometry padding;
-  final Color backgroundColor;
-  final TextStyle defaultTextStyle;
-  final OnLinkTap onLinkTap;
+  final EdgeInsetsGeometry? padding;
+  final Color? backgroundColor;
+  final TextStyle? defaultTextStyle;
+  final OnLinkTap? onLinkTap;
   final bool renderNewlines;
   final double blockSpacing;
   final bool useRichText;
-  final ImageErrorListener onImageError;
+  final ImageErrorListener? onImageError;
   final TextStyle linkStyle;
   final bool shrinkToFit;
 
   /// Properties for the Image widget that gets rendered by the rich text parser
-  final ImageProperties imageProperties;
-  final OnImageTap onImageTap;
+  final ImageProperties? imageProperties;
+  final OnImageTap? onImageTap;
   final bool showImages;
 
   /// Either return a custom widget for specific node types or return null to
   /// fallback to the default rendering.
-  final CustomRender customRender;
-  final CustomEdgeInsets customEdgeInsets;
-  final CustomTextStyle customTextStyle;
-  final CustomTextAlign customTextAlign;
+  final CustomRender? customRender;
+  final CustomEdgeInsets? customEdgeInsets;
+  final CustomTextStyle? customTextStyle;
+  final CustomTextAlign? customTextAlign;
 
   @override
   Widget build(BuildContext context) {
-    final double width = shrinkToFit ? null : MediaQuery.of(context).size.width;
+    final double? width = shrinkToFit ? null : MediaQuery.of(context).size.width;
 
     return Container(
       padding: padding,
